requirements.txt
aspy.yaml==1.3.0
bcrypt==3.1.7
cffi==1.12.3
cfgv==2.0.1
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-environ==0.4.5
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
identify==1.4.7
importlib-metadata==0.22
mccabe==0.6.1
more-itertools==7.2.0
mysqlclient==1.4.4
nodeenv==1.3.3
pre-commit==1.18.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-json-logger==0.1.11
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
virtualenv==16.7.5
zipp==0.6.0
